organized folders & moved github chart to anlaytics page
