+++ src/components/analytics/analytics.js
@@ -2,15 +2,20 @@ import React from 'react';
 import { Bar, Line, Pie, Doughnut, Radar, Polar } from 'react-chartjs-2';
 import { Card, CardActions, CardHeader, CardMedia, CardTitle, CardText } from 'material-ui/Card';
 import FlatButton from 'material-ui/FlatButton';
+import GithubSkills from '../profile/githubSkills';
 
 import JobStatus from './jobstatus';
 import Applications from './applications';
 
-const Analytics = ({ savedJobs }) => {
+const Analytics = (props) => {
   return (
     <div>
-      <JobStatus savedJobs={savedJobs} />
-      <Applications savedJobs={savedJobs} />
+      <JobStatus savedJobs={props.savedJobs} />
+      <Applications savedJobs={props.savedJobs} />
+      {Object.keys(props.githubSkills).length ?
+        <GithubSkills
+          githubSkills={props.githubSkills}
+        /> : null }
     </div>
   );
 };
